feat(rating): remove Axon dependency
