--- services/rating/src/main/java/com/eshop/rating/application/commandbus/Command.java
@@ -1,4 +1,0 @@
-package com.eshop.rating.application.commandbus;
-
-public interface Command<R> {
-}
