--- services/rating/src/main/java/com/eshop/rating/application/commandbus/RatingCommandBus.java
@@ -1,5 +1,0 @@
-package com.eshop.rating.application.commandbus;
-
-public interface RatingCommandBus {
-  <R, C extends Command<R>> R execute(C command);
-}
