--- services/rating/src/main/java/com/eshop/rating/application/commandbus/RatingCommandHandler.java
@@ -1,5 +1,0 @@
-package com.eshop.rating.application.commandbus;
-
-public interface RatingCommandHandler<R, C extends Command<R>> {
-  R handle(C command);
-}
