+++ services/rating/src/main/java/com/eshop/rating/application/commands/AddRatingCommandHandler.java
@@ -1,13 +1,12 @@
 package com.eshop.rating.application.commands;
 
-import com.eshop.rating.application.commandbus.RatingCommandHandler;
+import com.eshop.rating.application.shared.CommandHandler;
 import com.eshop.rating.application.model.RatingForCatalogItemDto;
 import com.eshop.rating.application.services.CalculateRatingForCatalogItemService;
 import com.eshop.rating.model.Rating;
 import com.eshop.rating.model.RatingRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.axonframework.commandhandling.CommandHandler;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,14 +15,13 @@ import java.util.UUID;
 @RequiredArgsConstructor
 @Component
 @Slf4j
-public class AddRatingCommandHandler implements RatingCommandHandler<RatingForCatalogItemDto, AddRatingCommand> {
+public class AddRatingCommandHandler implements CommandHandler<RatingForCatalogItemDto, AddRatingCommand> {
 
   private final CalculateRatingForCatalogItemService calculateRatingForCatalogItemService;
   private final RatingRepository ratingRepository;
 
   @Override
   @Transactional
-  @CommandHandler
   public RatingForCatalogItemDto handle(AddRatingCommand command) {
     var newRating = Rating.builder()
       .id(UUID.randomUUID())
