+++ services/rating/src/main/java/com/eshop/rating/application/queries/ratingforcatalogitem/RatingForCatalogItemQuery.java
@@ -1,7 +1,7 @@
 package com.eshop.rating.application.queries.ratingforcatalogitem;
 
 import com.eshop.rating.application.model.RatingForCatalogItemDto;
-import com.eshop.rating.application.querybus.Query;
+import com.eshop.rating.application.shared.Query;
 
 import java.util.UUID;
