+++ services/rating/src/main/java/com/eshop/rating/application/queries/ratingforcatalogitem/RatingForCatalogItemQueryHandler.java
@@ -1,7 +1,7 @@
 package com.eshop.rating.application.queries.ratingforcatalogitem;
 
 import com.eshop.rating.application.model.RatingForCatalogItemDto;
-import com.eshop.rating.application.querybus.QueryHandler;
+import com.eshop.rating.application.shared.QueryHandler;
 import com.eshop.rating.application.services.CalculateRatingForCatalogItemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
