+++ services/rating/src/main/java/com/eshop/rating/application/queries/ratingsbycatalogitem/RatingsByCatalogItemIdQuery.java
@@ -1,6 +1,6 @@
 package com.eshop.rating.application.queries.ratingsbycatalogitem;
 
-import com.eshop.rating.application.querybus.Query;
+import com.eshop.rating.application.shared.Query;
 import com.eshop.rating.model.Rating;
 
 import java.util.UUID;
