--- services/rating/src/main/java/com/eshop/rating/application/querybus/Query.java
@@ -1,4 +1,0 @@
-package com.eshop.rating.application.querybus;
-
-public interface Query<R> {
-}
