+++ services/rating/src/main/java/com/eshop/rating/application/shared/CommandBus.java
@@ -1,0 +1,5 @@
+package com.eshop.rating.application.shared;
+
+public interface CommandBus {
+  <R, C extends Command<R>> R execute(C command);
+}
