+++ services/rating/src/main/java/com/eshop/rating/application/shared/CommandHandler.java
@@ -1,0 +1,11 @@
+package com.eshop.rating.application.shared;
+
+public interface CommandHandler<R, C extends Command<R>> {
+  R handle(C command);
+
+  default boolean matches(C command) {
+    final var handlerType = getClass();
+    final var commandType = command.getClass();
+    return new FirstGenericArgOf(handlerType).isAssignableFrom(commandType);
+  }
+}
