+++ services/rating/src/main/java/com/eshop/rating/application/shared/Query.java
@@ -1,0 +1,4 @@
+package com.eshop.rating.application.shared;
+
+public interface Query<R> {
+}
