+++ services/rating/src/main/java/com/eshop/rating/application/shared/QueryBus.java
@@ -1,4 +1,4 @@
-package com.eshop.rating.application.querybus;
+package com.eshop.rating.application.shared;
 
 public interface QueryBus {
   <R, Q extends Query<R>> R execute(Q query);
